services/python-flask/requirements.txt
Flask==3.1.0
PyNaCl==1.5.0
google-cloud-pubsub==2.26.1
